refactor: move LoaderGLTF to ModelImpl
