+++ src/scene/CMakeLists.txt
@@ -5,5 +5,4 @@ target_sources(${PROJECT_NAME}
         Node.cpp
         ParentSystem.cpp
         TransformSystem.cpp
-        LoaderGLTF.cpp
 )
